requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
asgiref==3.10.0
astunparse==1.6.3
boto3==1.40.70
botocore==1.40.70
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.7
django-storages==1.14.6
flatbuffers==25.9.23
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h5py==3.15.1
httplib2==0.31.0
idna==3.11
jmespath==1.0.1
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.3.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rsa==4.9.1
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
